micu/verify_service/requirements.txt
Flask==2.3.2
pillow==10.0.0
numpy==1.26.2
pytesseract==0.3.10
opencv-python==4.8.1.78
torch==2.2.0
werkzeug==3.0.0
torch==2.2.0
werkzeug==3.0.0
